feat: add MythicScribe.openLogs command to open the extension's logs for debugging

+++ src/datasets/customDatasets.ts
@@ -5,7 +5,7 @@ import * as vscode from 'vscode';
 import { MechanicDataset, ScribeMechanicHandler } from './ScribeMechanic';
 import { ScribeEnumHandler, StaticScribeEnum, WebScribeEnum } from './ScribeEnum';
 import { fetchMechanicDatasetFromLink, loadDatasets, loadLocalMechanicDataset } from './datasets';
-import { logError, logInfo } from '../utils/logger';
+import { logDebug, logError, logInfo } from '../utils/logger';
 import { changeCustomDatasetsSource } from '../migration/migration';
 
 enum CustomDatasetElementType {
@@ -34,6 +34,7 @@ const validConfigurationTargets = [
 ];
 
 export async function addCustomDataset() {
+    logDebug('addCustomDataset');
     const scope = await vscode.window
         .showQuickPick(validConfigurationTargets, {
             placeHolder: 'Select the scope for which you want to add the custom dataset',
@@ -43,6 +44,7 @@ export async function addCustomDataset() {
     if (!scope) {
         return;
     }
+    logDebug('addCustomDataset scope:', scope.toString());
 
     const elementType = await vscode.window.showQuickPick(Object.values(CustomDatasetElementType), {
         placeHolder: 'Select an element type',
@@ -51,6 +53,7 @@ export async function addCustomDataset() {
     if (!elementType) {
         return;
     }
+    logDebug('addCustomDataset elementType:', elementType);
 
     const source = await vscode.window.showQuickPick(Object.values(CustomDatasetSource), {
         placeHolder: 'Select a source',
@@ -59,6 +62,7 @@ export async function addCustomDataset() {
     if (!source) {
         return;
     }
+    logDebug('addCustomDataset source:', source);
 
     if (source === CustomDatasetSource.LINK) {
         return addCustomDatasetFromLink(elementType, scope);
@@ -76,6 +80,8 @@ export async function addCustomDataset() {
         logInfo('No file selected.');
         return;
     }
+
+    logDebug('addCustomDataset fileUri:', fileUri.join(', '));
 
     //const mechanicDatasets : MechanicDataset[] = [];
     const validPaths: vscode.Uri[] = [];
@@ -116,9 +122,16 @@ export async function addCustomDataset() {
 }
 
 export async function removeCustomDataset() {
+    logDebug('removeCustomDataset');
     const scope = await vscode.window.showQuickPick(validConfigurationTargets, {
         placeHolder: 'Select the scope from which you want to remove the custom dataset',
     });
+
+    if (!scope) {
+        return;
+    }
+
+    logDebug('removeCustomDataset scope:', scope.target.toString());
 
     const [config, existingMappings] = getCustomDatasetConfiguration(scope?.target);
 
@@ -140,6 +153,11 @@ export async function removeCustomDataset() {
         return;
     }
 
+    logDebug(
+        'removeCustomDataset datasets:',
+        datasets.map((dataset) => dataset.pathOrUrl).join(', ')
+    );
+
     datasets.forEach(async (dataset) => {
         const index = existingMappings.indexOf(dataset);
         existingMappings.splice(index, 1);
@@ -154,9 +172,16 @@ export async function removeCustomDataset() {
 }
 
 export async function createBundleDataset() {
+    logDebug('createBundleDataset');
     const scope = await vscode.window.showQuickPick(validConfigurationTargets, {
         placeHolder: 'Select the scope for which you want to create the bundle',
     });
+
+    if (!scope) {
+        return;
+    }
+
+    logDebug('createBundleDataset scope:', scope.target.toString());
 
     const [config, existingMappings] = getCustomDatasetConfiguration(scope?.target);
 
@@ -186,6 +211,8 @@ export async function createBundleDataset() {
         return;
     }
 
+    logDebug('createBundleDataset bundlePath:', bundlePath.fsPath);
+
     const bundleData = datasets.map((dataset) => {
         if (dataset.mapping.source === CustomDatasetSource.FILE) {
             const bundleDir = path.dirname(bundlePath.fsPath);
@@ -208,6 +235,15 @@ export async function createBundleDataset() {
         }
     );
 
+    if (shoudReplaceSelectedCustomDatasetsWithBundle === undefined) {
+        return;
+    }
+
+    logDebug(
+        'createBundleDataset shoudReplaceSelectedCustomDatasetsWithBundle:',
+        shoudReplaceSelectedCustomDatasetsWithBundle.value.toString()
+    );
+
     try {
         await vscode.workspace.fs.writeFile(
             bundlePath,
@@ -219,7 +255,7 @@ export async function createBundleDataset() {
         return;
     }
 
-    if (shoudReplaceSelectedCustomDatasetsWithBundle?.value) {
+    if (shoudReplaceSelectedCustomDatasetsWithBundle.value) {
         for (const dataset of datasets) {
             const index = existingMappings.indexOf(dataset.mapping);
             if (index !== -1) {
@@ -304,9 +340,11 @@ function getCustomDatasetConfiguration(
 }
 
 export async function loadCustomDatasets() {
+    logDebug('loading custom datasets');
     const customDatasets = getCustomDatasetConfiguration()[1];
     for (const entry of customDatasets) {
         if (isOutdatedDataset(entry)) {
+            logDebug('Outdated dataset to migrate found:', entry.pathOrUrl);
             await changeCustomDatasetsSource('customDatasets', 'Local File', 'File');
             break;
         }
@@ -317,6 +355,7 @@ export async function loadCustomDatasets() {
 }
 
 async function processCustomDatasetEntry(entry: CustomDataset) {
+    logDebug('Processing custom dataset entry:', entry.pathOrUrl);
     if (entry.elementType === CustomDatasetElementType.BUNDLE) {
         await loadBundleDataset(entry);
     } else if (entry.elementType === CustomDatasetElementType.ENUM) {
@@ -352,6 +391,7 @@ async function processMechanicDatasetEntry(
 }
 
 async function loadBundleDataset(dataset: CustomDataset) {
+    logDebug('Loading bundle dataset:', dataset.pathOrUrl);
     try {
         const configData: CustomDataset[] = [];
         const fileUri = vscode.Uri.parse(dataset.pathOrUrl);
@@ -362,10 +402,12 @@ async function loadBundleDataset(dataset: CustomDataset) {
                 ...((JSON.parse(Buffer.from(fileContent).toString('utf8')) as CustomDataset[]) ||
                     [])
             );
+            logDebug('Bundle dataset loaded as file:', dataset.pathOrUrl);
         } else if (isLinkSource(dataset.source)) {
             const response = await fetch(dataset.pathOrUrl);
             const jsonData = (await response.json()) as CustomDataset[];
             configData.push(...jsonData);
+            logDebug('Bundle dataset loaded as link:', dataset.pathOrUrl);
         }
 
         for (const entry of configData) {
@@ -376,6 +418,7 @@ async function loadBundleDataset(dataset: CustomDataset) {
                 const entryUri = vscode.Uri.joinPath(datasetDirUri, entry.pathOrUrl);
                 entry.pathOrUrl = entryUri.toString();
                 entry.source = dataset.source;
+                logDebug('Resolved relative path:', entry.pathOrUrl);
             }
             await processCustomDatasetEntry(entry);
         }
